css/cssom-view/scrollIntoView-flipped-writing-mode.html

<!DOCTYPE html>
<title>CSSOM View - scrollIntoView considers flipped writing mode</title>
<meta charset="utf-8">
<link rel="author" title="Suneel Kota" href="mailto:suneel.kota@samsung.com">
<link rel="help" href="https://drafts.csswg.org/cssom-view/#dom-element-scrollintoview">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
#scroller {
  writing-mode: vertical-rl;
  overflow: hidden;
  width: 500px;
  height: 500px;
}
.filling {
  width: 500px;
}
#target {
  background-color: #ff0;
  width: 50px;
  height: 100px;
}
</style>
<body>
<div id="scroller">
  <div class="filling"></div>
  <div id="target"></div>
  <div class="filling"></div>
</div>
</div>
<script>
var target = document.getElementById("target");
var scroller = document.getElementById("scroller");

var filling = 500;

var expectedXCenter = (filling/2) + (target.offsetWidth/2);
var expectedXStart = target.offsetWidth;
var expectedXEnd = filling;

// This formats dict as a string suitable as test name.
// format_value() is provided by testharness.js,
// which also preserves sign for -0.
function format_dict(dict) {
  const props = [];
  for (let prop in dict) {
    props.push(`${prop}: ${format_value(dict[prop])}`);
  }
  return `{${props.join(", ")}}`;
}

[
  [{block: "center", inline: "center"}, expectedXCenter],
  [{block: "start", inline: "start"}, expectedXStart],
  [{block: "end", inline: "end"}, expectedXEnd],
].forEach(([input, expectedX]) => {
  test(() => {
    scroller.scrollTo(0, 0);
    target.scrollIntoView(input);
    assert_approx_equals(scroller.scrollLeft, expectedX, 0.5, "scrollX");
  }, `scrollIntoView(${format_dict(input)})`);
})
</script>
